requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dataclasses==0.8
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fuzzywuzzy==0.18.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
numpy==1.19.5
pandas==1.1.5
pkg-resources==0.0.0
protobuf==3.19.1
pyenchant==3.2.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
spatialite==0.0.3
threadpoolctl==3.0.0
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
